DOC: rename Developer Guide to Core Developer Guide
New SciPy contributors are likely to think of themselves as
"Developers", but this guide is geared toward Core Developers.
The name made it sound like it was a good place to start learning
about developing for SciPy, but it isn't. I hope this slight change
helps to distinguish it from the rest of the contributor documentation.