MAINT: stats.ContinuousDistribution: fix drawing variable from 'typical'
Previously, the `draw` method of `_Parameter` relied entirely on either
the `draw` method of its `domain` or the `draw` method of its `typical`
attribute. Consequently, when drawing a "typical" value of the
parameter, endpoint values would be the endpoint of the *typical* values;
outside values would be outside the typical values but possibly still
inside the greater domain. This restructures the `draw` method such that
`_Parameter` can draw using the `draw` method of multiple subdomain
attributes.